+++ library/test/src/test/java/com/bumptech/glide/manager/DefaultConnectivityMonitorFactoryTest.java
@@ -31,10 +31,42 @@ public class DefaultConnectivityMonitorFactoryTest {
   }
 
   @Test
-  public void testReturnsNullConnectivityMonitorWhenDoesNotHavePermission() {
+  @Config(sdk = 18)
+  public void testDefaultConnectivityMonitorInjectPreMConnectivityStrategyImplWhenHasPermission() {
+    ShadowApplication.getInstance().grantPermissions("android.permission.ACCESS_NETWORK_STATE");
     ConnectivityMonitor connectivityMonitor =
         factory.build(
             RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
-    assertThat(connectivityMonitor).isInstanceOf(NullConnectivityMonitor.class);
+    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
+    DefaultConnectivityMonitor defaultConnectivityMonitor =
+        (DefaultConnectivityMonitor) connectivityMonitor;
+    assertThat(defaultConnectivityMonitor.strategy)
+        .isInstanceOf(PreMConnectivityStrategyImpl.class);
+  }
+
+  @Test
+  @Config(sdk = 23)
+  public void testDefaultConnectivityMonitorInjectPostMConnectivityStrategyImplWhenHasPermission() {
+    ShadowApplication.getInstance().grantPermissions("android.permission.ACCESS_NETWORK_STATE");
+    ConnectivityMonitor connectivityMonitor =
+        factory.build(
+            RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
+    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
+    DefaultConnectivityMonitor defaultConnectivityMonitor =
+        (DefaultConnectivityMonitor) connectivityMonitor;
+    assertThat(defaultConnectivityMonitor.strategy)
+        .isInstanceOf(PostMConnectivityStrategyImpl.class);
+  }
+
+  @Test
+  public void testDefaultConnectivityMonitorInjectNullConnectivityStrategyImplNotHavePermission() {
+    ConnectivityMonitor connectivityMonitor =
+        factory.build(
+            RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
+    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
+    DefaultConnectivityMonitor defaultConnectivityMonitor =
+        (DefaultConnectivityMonitor) connectivityMonitor;
+    assertThat(defaultConnectivityMonitor.strategy)
+        .isInstanceOf(NullConnectivityStrategyImpl.class);
   }
 }
